port the "xpos" API to [0.0..1.0] doubles too.
2008-02-09  Michael Natterer  <mitch@gimp.org>

	* app/widgets/gimpcurveview.[ch]: port the "xpos" API
	to [0.0..1.0] doubles too.

	* app/tools/gimpcurvestool.[ch]: rename "col_value" member to
	"picked_color" and use gdouble instead of gint. Also use GimpCurve
	API to map the values instead of accessing the curve directly.
	Fixes setting curve anchor points by color picking.


svn path=/trunk/; revision=24838

+++ app/widgets/gimpcurveview.h
@@ -52,7 +52,7 @@ struct _GimpCurveView
 
   GdkCursorType      cursor_type;
 
-  gint               xpos;
+  gdouble            xpos;
   PangoLayout       *xpos_layout;
 
   gint               cursor_x;
@@ -78,7 +78,7 @@ GimpCurve * gimp_curve_view_get_curve    (GimpCurveView *view);
 void        gimp_curve_view_set_selected (GimpCurveView *view,
                                           gint           selected);
 void        gimp_curve_view_set_xpos     (GimpCurveView *view,
-                                          gint           x);
+                                          gdouble        x);
 
 
 #endif /* __GIMP_CURVE_VIEW_H__ */
